+++ docs/changelog.md
@@ -5,7 +5,7 @@
 This release is a major overhaul of the tool including built-in download functionality.
 
 - Moved the extraction functionality to the `wpextract extract` subcommand
-- Integrate a heavily modified version of [WPJsonScraper](`https://github.com/MickaelWalter/wp-json-scraper`) as the `wpextract dl` subcommand
+- Integrate a heavily modified version of [WPJsonScraper](`https://github.com/MickaelWalter/wp-json-scraper`) as the `wpextract download` subcommand
 - Renamed the main package of this library to `wpextract` to match the CLI tool name
 - Support extraction without an HTML scrape if translations aren't needed
 - Support extracting only some of the possible data types